backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.75.0
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3
yfinance==0.2.55
requests_cache==1.2.1
requests_ratelimiter==0.7.0
pyrate-limiter==2.10.0
openai-agents==0.1.1
python-dateutil==2.9.0 